Move wrappers and their stubs into ext folder
