+++ ext/log/logtest/stubs.go
