+++ ext/viper/vipertest/stubs.go
