+++ login/login_test.go
@@ -3,21 +3,13 @@ package login
 import (
 	"errors"
 	"fmt"
+	httpService "github.com/betasve/mstd/ext/http/httptest"
+	"github.com/betasve/mstd/ext/time"
 	"io/ioutil"
+	"net/http"
+	"net/url"
 	"strconv"
 	"strings"
-	// "github.com/betasve/mstd/exec"
-	// exectest "github.com/betasve/mstd/exec/exectest"
-	// "github.com/betasve/mstd/log"
-	// logtest "github.com/betasve/mstd/log/logtest"
-	// "github.com/betasve/mstd/runtime"
-	// runtimetest "github.com/betasve/mstd/runtime/runtimetest"
-	"github.com/betasve/mstd/time"
-	"net/url"
-	// timetest "github.com/betasve/mstd/time/timetest"
-	// osexec "os/exec"
-	httpService "github.com/betasve/mstd/http/httptest"
-	"net/http"
 	"testing"
 	t "time"
 )
